# Conflicts:
#	module-1/src/MathematicalOperationsTest.java

+++ module-1/src/MathematicalOperations.java
@@ -60,7 +60,7 @@ public class MathematicalOperations {
     }
 
     /**
-     * Returns the pow power of base
+     * Returns the pow power of the base
      * Added by Oyku Yilmaz
      *
      * @param base integer1
@@ -90,10 +90,10 @@ public class MathematicalOperations {
     }
 
     /**
-     * Takes 2 integer to find inverse divide
+     * Takes 2 int to find inverse divide
      * Added by Ugur Tombul
     */
-    public static double InverseDivide(double a,double b) {
+    public static double InverseDivide(int a,int b) {
         if(a == 0){
             throw new ArithmeticException("divided by 0");
         }
